cleanup and remove transitional files. Load script now works with .ddb and writes partitioned files. Query script queries partitioned files
