--- test_roi.geojson
@@ -1,30 +1,0 @@
-{
-  "type": "FeatureCollection",
-  "features": [
-    {
-      "type": "Feature",
-      "properties": {
-        "name": "Test ROI from query_partitioned_tables.sql",
-        "description": "Bounding box: -85.605165, 30.357851, -80.839729, 35.000659"
-      },
-      "geometry": {
-        "type": "Polygon",
-        "coordinates": [
-          [
-            [-85.605165, 30.357851],
-            [-80.839729, 30.357851],
-            [-80.839729, 35.000659],
-            [-85.605165, 35.000659],
-            [-85.605165, 30.357851]
-          ]
-        ]
-      }
-    }
-  ],
-  "crs": {
-    "type": "name",
-    "properties": {
-      "name": "EPSG:4326"
-    }
-  }
-}
